Update brute_force.jl
the board array is allocated and cleared in every iteration of the inner loop. Instead, we can allocate it once before the loop and reuse it

In the brute_force and num_solutions functions, the combinations iterator is created in every iteration of the outer loop. It is more efficient to preallocate the combinations outside the loop and reuse it.